重构ssh

+++ cdh-5.9.0/ssh/fabfile2.py
@@ -2,6 +2,10 @@
 from __future__ import with_statement
 from fabric.api import run, env, execute, roles, runs_once, parallel, sudo, hosts, cd, task
 from fabric.contrib import files
+"""
+通过execute重复调用一个方法
+直接调用方法　区别
+"""
 
 # 通过roles与横块化来实现可重用
 # known_hosts处理，不然会提示信息
@@ -27,22 +31,25 @@ def append_authrozied_key(key):
 @runs_once
 @task
 def add_public_key_authorizied_keys(add_keys=[]):
-    # TODO 添加参数，实现额处添加key
+    """
+    添加参数，实现额处添加key
+    :param add_keys: 额处添加的key值
+    :return:
+    """
     execute(gen_public_key)
     keys = execute(get_id_ras_public_key)
-    local_key = """ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAABAQDHJXAZx0EzfFtyoFJvimiP1J4/XZZU7WOtx/cH2/dHAi5eBjqk5WjSRVgYGmJWCrivE2KEub25fGqR1wVnZXSfP3x9PsQnE5eO5I/K2OAfUGakHVZeYKm3J5ADryxcRxcKj8d0Y+/PHSyrL0smN+ZyyYqFJ2nfW2tGYKAK9bwdwjcf/QgnR558SbWAjKmBp81JUxQUpkNO+Hv4yIaPPbsYXsVk752DjIrgsE2riGrjCLju1yBu6T7mYw0wEpn4XrE0WQfyZkBRV/BS4Iz1iFYQExrsWaJbe82Gt00TJksbJJN81FGtllmo1AdGTVonwpv3dAS6AMfT2RcFrPZw018v congsl@congsl"""
-    keys['local'] = local_key
-    for host, key in keys.items():
+    vals = keys.values() + add_keys
+    for key in vals:
         execute(append_authrozied_key, key)
-        # execute(append_authrozied_key,local_key)
-        # append_authrozied_key(local_key)
 
 
 @task
 def clear_id_rsa():
-    run('rm ~/.ssh/id_rsa*')
-    run('rm ~/.ssh/authorized_keys')
-    run('rm ~/.ssh/known_hosts')
+    if files.exists('.ssh/id_rsa'):
+        run('rm ~/.ssh/id_rsa*')
+    if files.exists('.ssh/authorized_keys'):
+        run('rm ~/.ssh/authorized_keys')
+        # run('rm ~/.ssh/known_hosts')
 
 
 def test_hosts():
